+++ src/Order/Changelog.md
@@ -5,6 +5,8 @@
 ## Unreleased
 ##### 2024-XX-YY
 
+- Added `domain` field to the orders table
+- Added then automatic completion of `$order->domain` from the current request to the OrderFactory
 - Changed `OrderItem::hasConfiguration()` to return false on empty arrays as well
 
 ## 4.1.0